Shift arm screen to dorsum
