+++ Assets/Materials/Hand_Panel_Material.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 8f425f01e4f4d0a4789f98ba5e9e418f
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
